[UI] show login dialog in submit link activity if necessary
This took a lot longer than it should have (read: took a couple hours)
because turns out I was hitting a probable race condition when accessing
the SubmitLinkActivity from the Browser. I'm assuming because there was
less overhead so things happened differently. The problem was (I think)
that the login dialog was being dismissed before the username and password
got read. It was a simple fix, just save the user and pass to variables,
then dismiss dialog, then start new LoginTask. Ugh. Sounds like an obvious
thing to do, in retrospect.