progress: use the Window progressbar (in title bar)
Use a special FilterInputStream to update progress while parsing JSON.
Gotten from stackoverflow.
Using the number of bytes caused the android ProgressDialog to show
the bytesRead/totalBytes ratio which was confusing to me as a user.
So I replaced the ProgressDialogs.

The Window progress bar replaces the old ProgressDialog bars when
loading lists of threads and comments.

TODO: Make a nice loading screen while loading, since now it is just
a blank space with small progress bar in the titlebar.