cache: cache subreddit JSON. do not bother with Serializable
Seems to work pretty well. Definitely feels cleaner and more
predictable than before. Handle caching within the DownloadThreadsTask
instead of in a separate ReadCacheTask. And do not bother saving
from saveInstanceState; just save the JSON immediately when you
read it from the server.

Still needs a lot of tweaking though. When to clear cache?
1. when refreshing
2. logging in/out
3. when interacting with the site (voting, commenting)

So cache kind of gets in the way. a lot of extra one-liners will
be added to delete the cache in all the different situations...