browser: use a transparent background while loading. SEE COMMIT LOG
Does not work on all sites, some of them have dark bg but it does not
kick in until after the flash of white bg.
At least it seems to work fine on imgur, where the dark imgur bg
will show up before the white bg behind it.

Example problematic site:
http://m.gametrailers.com

I think the difference may just be in loading speed?
Didn't see any JavaScript on gametrailers like deciding bgcolor
on the fly. bgcolor is in the <body> tag already.