Added r/all to the list of subreddits presented in the subreddit picker, as requested here:
https://github.com/talklittle/reddit-is-fun/issues/116

In addition, I added support for the other "Fake" subreddits such as r/random and r/friends, they could be added to the FAKE_SUBREDDITS array. r/all is a much needed feature, because if you want to see the logged out reddit frontpage, you had to manually type in all each time. Users without specific subreddit subscriptions might not understand this change, but it was quite obnoxious for some users.